refactor[#56]: merge

+++ backend/src/main/java/com/handong/rebon/review/domain/content/ReviewScore.java
@@ -21,7 +21,7 @@ public class ReviewScore {
         this.empathyCount = empathyCount;
     }
 
-    private void validatesRangeOfValues(double star, int likeCount) {
+    private void validatesRangeOfValues(int star, int likeCount) {
         if (star <= 0 || star > 5) {
             throw new ReviewStarException();
         }
